feat: protect routes

+++ timemanager-api/src/main/java/edu/epitech/timemanager/api/UserController.java
@@ -10,6 +10,7 @@ import lombok.extern.slf4j.Slf4j;
 import org.mapstruct.factory.Mappers;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
+import org.springframework.security.core.annotation.AuthenticationPrincipal;
 import org.springframework.web.bind.annotation.*;
 
 
@@ -20,6 +21,11 @@ import org.springframework.web.bind.annotation.*;
 public class UserController {
     private final UserService userService;
     private final UserMappers userMappers = Mappers.getMapper(UserMappers.class);
+
+    @GetMapping("")
+    public ResponseEntity<?> getUsers() {
+        return ResponseEntity.ok(userService.getUsers());
+    }
 
     @GetMapping("/{id}")
     public ResponseEntity<?> getUser(@PathVariable("id") int id) {
@@ -66,7 +72,15 @@ public class UserController {
     }
 
     @PutMapping("/{id}")
-    public ResponseEntity<?> updateUser(@PathVariable("id") int id, @RequestBody UpdateUserDto userDto) {
+    public ResponseEntity<?> updateUser(
+        @AuthenticationPrincipal User authenticateUser,
+        @PathVariable("id") int id,
+        @RequestBody UpdateUserDto userDto
+    ) {
+        if (authenticateUser.getId() != id) {
+            return new ResponseEntity<>(HttpStatus.UNAUTHORIZED);
+        }
+
         User user = userService.updateUser(id, new User(
             id,
             userDto.getUsername(),
@@ -81,7 +95,14 @@ public class UserController {
     }
 
     @DeleteMapping("/{id}")
-    public ResponseEntity<?> deleteUser(@PathVariable("id") int id) {
+    public ResponseEntity<?> deleteUser(
+        @AuthenticationPrincipal User authenticateUser,
+        @PathVariable("id") int id
+    ) {
+        if (authenticateUser.getId() != id) {
+            return new ResponseEntity<>(HttpStatus.UNAUTHORIZED);
+        }
+
         userService.deleteUser(id);
         return new ResponseEntity<>(HttpStatus.OK);
     }
